+++ json-providers/gson-json-provider/src/main/java/ru/easydonate/easydonate4j/api/v3/data/model/gson/shop/purchase/PurchasedProductModel.java
@@ -34,9 +34,9 @@ public class PurchasedProductModel implements PurchasedProduct {
     @SerializedName("price")
     private double price;
     @SerializedName("number")
-    private int defaultAmount;
+    private long defaultAmount;
     @SerializedName("amount")
-    private int boughtAmount;
+    private long boughtAmount;
 
     @SerializedName("description")
     private String description;
