+++ json-providers/gson-json-provider/src/main/java/ru/easydonate/easydonate4j/callback/data/model/gson/purchase/PurchasedProductModel.java
@@ -25,7 +25,7 @@ public class PurchasedProductModel implements PurchasedProduct {
     @SerializedName("cost")
     private double price;
     @SerializedName("count")
-    private int boughtAmount;
+    private long boughtAmount;
 
     @SerializedName("description")
     private String description;
